Rename debug.* files to debug_logger.* - seemed to conflict with make debug target.
